Detect deliver mission
