+++ mission_control/navigator_missions/navigator_missions/__init__.py
@@ -5,7 +5,6 @@ ChainWithTimeout = mil_tasks_core.MakeChainWithTimeout(Navigator)
 Wait = mil_tasks_core.MakeWait(Navigator)
 del mil_tasks_core
 
-from detect_deliver import DetectDeliver
 from teleop import Teleop
 from circle import Circle
 from circle_tower import CircleTower
@@ -34,4 +33,6 @@ from obstacle_avoid import ObstacleAvoid
 from stc_jaxon import ScanTheCodeJaxon
 from explore_towers import ExploreTowers
 from track_target import TrackTarget
+from detect_deliver_find import DetectDeliverFind
+from detect_deliver import DetectDeliver
 import pose_editor
